+++ instrumentation-genai/opentelemetry-instrumentation-vertexai/tests/test_function_calling_experimental.py
@@ -274,6 +274,14 @@ def test_tool_events_with_completion_hook(
     assert len(logs) == 1
     # File upload takes a few seconds sometimes.
     time.sleep(3)
+
+    # Both log and span have the same reference attributes from upload hook
+    for key in "gen_ai.input.messages_ref", "gen_ai.output.messages_ref":
+        assert spans[0].attributes.get(key)
+        assert logs[0].log_record.attributes.get(key)
+
+        assert spans[0].attributes[key] == logs[0].log_record.attributes[key]
+
     assert_fsspec_equal(
         spans[0].attributes["gen_ai.output.messages_ref"],
         [