+++ instrumentation-genai/opentelemetry-instrumentation-vertexai/src/opentelemetry/instrumentation/vertexai/__init__.py
@@ -47,7 +47,7 @@ from wrapt import (
     wrap_function_wrapper,  # type: ignore[reportUnknownVariableType]
 )
 
-from opentelemetry._events import get_event_logger
+from opentelemetry._logs import get_logger
 from opentelemetry.instrumentation._semconv import (
     _OpenTelemetrySemanticConventionStability,
     _OpenTelemetryStabilitySignalType,
@@ -60,6 +60,7 @@ from opentelemetry.instrumentation.vertexai.patch import MethodWrappers
 from opentelemetry.instrumentation.vertexai.utils import is_content_enabled
 from opentelemetry.semconv.schemas import Schemas
 from opentelemetry.trace import get_tracer
+from opentelemetry.util.genai.completion_hook import load_completion_hook
 
 
 def _methods_to_wrap(
@@ -109,6 +110,9 @@ class VertexAIInstrumentor(BaseInstrumentor):
 
     def _instrument(self, **kwargs: Any):
         """Enable VertexAI instrumentation."""
+        completion_hook = (
+            kwargs.get("completion_hook") or load_completion_hook()
+        )
         sem_conv_opt_in_mode = _OpenTelemetrySemanticConventionStability._get_opentelemetry_stability_opt_in_mode(
             _OpenTelemetryStabilitySignalType.GEN_AI,
         )
@@ -124,12 +128,12 @@ class VertexAIInstrumentor(BaseInstrumentor):
             tracer_provider,
             schema_url=schema,
         )
-        event_logger_provider = kwargs.get("event_logger_provider")
-        event_logger = get_event_logger(
+        logger_provider = kwargs.get("logger_provider")
+        logger = get_logger(
             __name__,
             "",
+            logger_provider=logger_provider,
             schema_url=schema,
-            event_logger_provider=event_logger_provider,
         )
         sem_conv_opt_in_mode = _OpenTelemetrySemanticConventionStability._get_opentelemetry_stability_opt_in_mode(
             _OpenTelemetryStabilitySignalType.GEN_AI,
@@ -138,17 +142,19 @@ class VertexAIInstrumentor(BaseInstrumentor):
             # Type checker now knows sem_conv_opt_in_mode is a Literal[_StabilityMode.DEFAULT]
             method_wrappers = MethodWrappers(
                 tracer,
-                event_logger,
+                logger,
                 is_content_enabled(sem_conv_opt_in_mode),
                 sem_conv_opt_in_mode,
+                completion_hook,
             )
         elif sem_conv_opt_in_mode == _StabilityMode.GEN_AI_LATEST_EXPERIMENTAL:
             # Type checker now knows it's the other literal
             method_wrappers = MethodWrappers(
                 tracer,
-                event_logger,
+                logger,
                 is_content_enabled(sem_conv_opt_in_mode),
                 sem_conv_opt_in_mode,
+                completion_hook,
             )
         else:
             raise RuntimeError(f"{sem_conv_opt_in_mode} mode not supported")