+++ instrumentation-genai/opentelemetry-instrumentation-vertexai/src/opentelemetry/instrumentation/vertexai/patch.py
@@ -15,6 +15,7 @@
 from __future__ import annotations
 
 from contextlib import contextmanager
+from dataclasses import asdict
 from typing import (
     TYPE_CHECKING,
     Any,
@@ -27,13 +28,14 @@ from typing import (
     overload,
 )
 
-from opentelemetry._events import EventLogger
+from opentelemetry._logs import Logger, LogRecord
 from opentelemetry.instrumentation._semconv import (
     _StabilityMode,
 )
 from opentelemetry.instrumentation.vertexai.utils import (
     GenerateContentParams,
-    create_operation_details_event,
+    _map_finish_reason,
+    convert_content_to_message_parts,
     get_genai_request_attributes,
     get_genai_response_attributes,
     get_server_attributes,
@@ -41,8 +43,17 @@ from opentelemetry.instrumentation.vertexai.utils import (
     request_to_events,
     response_to_events,
 )
+from opentelemetry.semconv._incubating.attributes import (
+    gen_ai_attributes as GenAI,
+)
 from opentelemetry.trace import SpanKind, Tracer
-from opentelemetry.util.genai.types import ContentCapturingMode
+from opentelemetry.util.genai.completion_hook import CompletionHook
+from opentelemetry.util.genai.types import (
+    ContentCapturingMode,
+    InputMessage,
+    OutputMessage,
+)
+from opentelemetry.util.genai.utils import gen_ai_json_dumps
 
 if TYPE_CHECKING:
     from google.cloud.aiplatform_v1.services.prediction_service import client
@@ -105,36 +116,40 @@ class MethodWrappers:
     def __init__(
         self,
         tracer: Tracer,
-        event_logger: EventLogger,
+        logger: Logger,
         capture_content: ContentCapturingMode,
         sem_conv_opt_in_mode: Literal[
             _StabilityMode.GEN_AI_LATEST_EXPERIMENTAL
         ],
+        completion_hook: CompletionHook,
     ) -> None: ...
 
     @overload
     def __init__(
         self,
         tracer: Tracer,
-        event_logger: EventLogger,
+        logger: Logger,
         capture_content: bool,
         sem_conv_opt_in_mode: Literal[_StabilityMode.DEFAULT],
+        completion_hook: CompletionHook,
     ) -> None: ...
 
     def __init__(
         self,
         tracer: Tracer,
-        event_logger: EventLogger,
+        logger: Logger,
         capture_content: Union[bool, ContentCapturingMode],
         sem_conv_opt_in_mode: Union[
             Literal[_StabilityMode.DEFAULT],
             Literal[_StabilityMode.GEN_AI_LATEST_EXPERIMENTAL],
         ],
+        completion_hook: CompletionHook,
     ) -> None:
         self.tracer = tracer
-        self.event_logger = event_logger
+        self.logger = logger
         self.capture_content = capture_content
         self.sem_conv_opt_in_mode = sem_conv_opt_in_mode
+        self.completion_hook = completion_hook
 
     @contextmanager
     def _with_new_instrumentation(
@@ -146,18 +161,10 @@ class MethodWrappers:
         kwargs: Any,
     ):
         params = _extract_params(*args, **kwargs)
-        api_endpoint: str = instance.api_endpoint  # type: ignore[reportUnknownMemberType]
-        span_attributes = {
-            **get_genai_request_attributes(False, params),
-            **get_server_attributes(api_endpoint),
-        }
-
-        span_name = get_span_name(span_attributes)
-
+        request_attributes = get_genai_request_attributes(True, params)
         with self.tracer.start_as_current_span(
-            name=span_name,
+            name=f"{GenAI.GenAiOperationNameValues.CHAT.value} {request_attributes.get(GenAI.GEN_AI_REQUEST_MODEL, '')}".strip(),
             kind=SpanKind.CLIENT,
-            attributes=span_attributes,
         ) as span:
 
             def handle_response(
@@ -165,22 +172,78 @@ class MethodWrappers:
                 | prediction_service_v1beta1.GenerateContentResponse
                 | None,
             ) -> None:
-                if span.is_recording() and response:
-                    # When streaming, this is called multiple times so attributes would be
-                    # overwritten. In practice, it looks the API only returns the interesting
-                    # attributes on the last streamed response. However, I couldn't find
-                    # documentation for this and setting attributes shouldn't be too expensive.
-                    span.set_attributes(
-                        get_genai_response_attributes(response)
-                    )
-                self.event_logger.emit(
-                    create_operation_details_event(
-                        api_endpoint=api_endpoint,
-                        params=params,
-                        capture_content=capture_content,
-                        response=response,
-                    )
+                attributes = (
+                    get_server_attributes(instance.api_endpoint)  # type: ignore[reportUnknownMemberType]
+                    | request_attributes
+                    | get_genai_response_attributes(response)
                 )
+                event = LogRecord(
+                    event_name="gen_ai.client.inference.operation.details",
+                )
+                event.attributes = attributes.copy()
+                system_instructions, inputs, outputs = [], [], []
+                if params.system_instruction:
+                    system_instructions = convert_content_to_message_parts(
+                        params.system_instruction
+                    )
+                if params.contents:
+                    inputs = [
+                        InputMessage(
+                            role=content.role,
+                            parts=convert_content_to_message_parts(content),
+                        )
+                        for content in params.contents
+                    ]
+                if response:
+                    outputs = [
+                        OutputMessage(
+                            finish_reason=_map_finish_reason(
+                                candidate.finish_reason
+                            ),
+                            role=candidate.content.role,
+                            parts=convert_content_to_message_parts(
+                                candidate.content
+                            ),
+                        )
+                        for candidate in response.candidates
+                    ]
+                self.completion_hook.on_completion(
+                    inputs=inputs,
+                    outputs=outputs,
+                    system_instruction=system_instructions,
+                    span=span,
+                    log_record=event,
+                )
+                content_attributes = {
+                    k: [asdict(x) for x in v]
+                    for k, v in [
+                        (
+                            GenAI.GEN_AI_SYSTEM_INSTRUCTIONS,
+                            system_instructions,
+                        ),
+                        (GenAI.GEN_AI_INPUT_MESSAGES, inputs),
+                        (GenAI.GEN_AI_OUTPUT_MESSAGES, outputs),
+                    ]
+                    if v
+                }
+                if span.is_recording():
+                    span.set_attributes(attributes)
+                    if capture_content in (
+                        ContentCapturingMode.SPAN_AND_EVENT,
+                        ContentCapturingMode.SPAN_ONLY,
+                    ):
+                        span.set_attributes(
+                            {
+                                k: gen_ai_json_dumps(v)
+                                for k, v in content_attributes.items()
+                            }
+                        )
+                if capture_content in (
+                    ContentCapturingMode.SPAN_AND_EVENT,
+                    ContentCapturingMode.EVENT_ONLY,
+                ):
+                    event.attributes |= content_attributes
+                self.logger.emit(event)
 
             yield handle_response
 
@@ -210,7 +273,7 @@ class MethodWrappers:
             for event in request_to_events(
                 params=params, capture_content=capture_content
             ):
-                self.event_logger.emit(event)
+                self.logger.emit(event)
 
             # TODO: set error.type attribute
             # https://github.com/open-telemetry/semantic-conventions/blob/main/docs/gen-ai/gen-ai-spans.md
@@ -231,7 +294,7 @@ class MethodWrappers:
                 for event in response_to_events(
                     response=response, capture_content=capture_content
                 ):
-                    self.event_logger.emit(event)
+                    self.logger.emit(event)
 
             yield handle_response