+++ instrumentation-genai/opentelemetry-instrumentation-vertexai/CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+- Fix overwritten log attributes in vertexai instrumentation
+  ([#3925](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/3925))
 
 ## Version 2.1b0 (2025-10-16)